v0.6.9 Updated accuracy and map calculation
Previously there was a bug in the calculation logic.
Instead of using misconception id when validating over all misconceptions,
labels were used which were incorrect. Labels represented the index of the
positive pair in the batch and not the misconception id. Now the actual
misconception ids are used for calculating accuracy and map@k.